+++ src/main/java/frc/robot/Robot.java
@@ -12,6 +12,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.commands.drivetrain.ResetGyro;
 import frc.robot.commands.drivetrain.WheelAlign;
+import frc.robot.constants.Constants;
 import frc.robot.utils.diag.Diagnostics;
 import frc.robot.utils.logging.Logger;
