+++ src/main/java/frc/robot/commands/cannon/ShootTest.java
@@ -2,7 +2,7 @@ package frc.robot.commands.cannon;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants;
+import frc.robot.constants.Constants;
 import frc.robot.subsystems.Shooter;
 
 public class ShootTest extends Command {
