+++ src/main/java/frc/robot/commands/climber/StaticClimb.java
@@ -1,7 +1,7 @@
 package frc.robot.commands.climber;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants;
+import frc.robot.constants.Constants;
 import frc.robot.subsystems.Climber;
 
 public class StaticClimb extends Command {
