+++ src/main/java/frc/robot/commands/drivetrain/WheelAlign.java
@@ -1,7 +1,7 @@
 package frc.robot.commands.drivetrain;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants;
+import frc.robot.constants.Constants;
 import frc.robot.subsystems.SwerveDrivetrain;
 
 public class WheelAlign extends Command {
