+++ src/main/java/frc/robot/subsystems/IntakeSubsystem.java
@@ -1,7 +1,7 @@
 package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
+import frc.robot.constants.Constants;
 import frc.robot.utils.smartshuffleboard.SmartShuffleboard;
 
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
